Wrote tests for the dishes endpoint DELETE GET POST UPDATE

+++ server/models/Dishes/Dishes.model.js
@@ -1,7 +1,12 @@
 const dishesCollection = require("./Dishes.mongo");
 async function getDishes(type, limit, skip) {
 	let totalCount = type === "all" ? await dishesCollection.countDocuments() : await dishesCollection.countDocuments({ type });
-	let response = type === "all" ? await dishesCollection.find({}, { __v: 0 }).limit(limit).skip(skip) : await dishesCollection.find({ type }, { __v: 0 }).limit(limit).skip(skip);
+	limit = limit || totalCount;
+	skip = skip || 0;
+	let response =
+		type === "all"
+			? await dishesCollection.find({}, { __v: 0 }).limit(limit).skip(skip).sort({ _id: -1 })
+			: await dishesCollection.find({ type }, { __v: 0 }).limit(limit).skip(skip).sort({ _id: -1 });
 	return { dishes: response, nextpage: parseInt(limit) + parseInt(skip) < totalCount };
 }
 
@@ -16,8 +21,10 @@ async function getADish(dishId) {
 }
 
 async function deleteADish(dishId) {
-	console.log(dishId);
 	try {
+		// Check to see if the chefs  exists
+		const exists = await dishesCollection.findOne({ _id: dishId });
+		if (!exists) throw new Error("No chef exists with the specified id");
 		await dishesCollection.deleteOne({ _id: dishId });
 	} catch (e) {
 		throw new Error("No dish exists with the specified id");
@@ -26,6 +33,9 @@ async function deleteADish(dishId) {
 
 async function updateADish(dishId, newData) {
 	try {
+		// Check to see if the chefs  exists
+		const exists = await dishesCollection.findOne({ _id: dishId });
+		if (!exists) throw new Error("No dish exists with the specified id");
 		await dishesCollection.updateOne({ _id: dishId }, newData);
 	} catch (e) {
 		throw new Error("No dish exists with the specified id");
